security patch

+++ lib/frontend/notifications.dart
@@ -67,21 +67,25 @@ class _NotificationsState extends State<Notifications> {
 
   void updatePage(BuildContext context, int index) {
     if (index == 0) {
-      print("Index: $index");
-      Firestore.instance
-          .collection("notifications")
-          .getDocuments()
-          .then((snapshot) {
-        for (var doc in snapshot.documents) {
-          doc.reference.delete();
-        }
-      }).catchError((onError){});
-      Firestore.instance
-          .collection("notifications")
-          .getDocuments()
-          .then((value) {
-        notifications = value.documents.length;
-      }).catchError((onError){});
+      if (isAdmin) {
+        Firestore.instance
+            .collection("notifications")
+            .getDocuments()
+            .then((snapshot) {
+          for (var doc in snapshot.documents) {
+            doc.reference.delete();
+          }
+        }).catchError((onError) {});
+        Firestore.instance
+            .collection("notifications")
+            .getDocuments()
+            .then((value) {
+          notifications = value.documents.length;
+        }).catchError((onError) {});
+      } else {
+        popUpInfo(context, "Alert",
+            "Error trying to delete notifications. Only an Administrator can delete notifications");
+      }
     }
     if (index == 1) {
       messageIcon = Icons.mail;
@@ -131,13 +135,6 @@ class _NotificationsState extends State<Notifications> {
                 Row(
                   mainAxisAlignment: MainAxisAlignment.spaceAround,
                   children: <Widget>[
-                    // Chip(
-                    //   avatar: CircleAvatar(
-                    //     backgroundColor: Colors.grey.shade800,
-                    //     child: Text(''),
-                    //   ),
-                    //   // label: Text(record.projectForeman),
-                    // ),
                     ButtonTheme.bar(
                       child: ButtonBar(
                         children: <Widget>[
@@ -150,17 +147,21 @@ class _NotificationsState extends State<Notifications> {
                               ),
                             ),
                             onPressed: () {
+                              if (isAdmin){
                               Firestore.instance
                                   .collection("notifications")
                                   .document(data.documentID)
                                   .delete();
-                              
+
                               Firestore.instance
                                   .collection("notifications")
                                   .getDocuments()
                                   .then((value) {
                                 notifications = value.documents.length;
                               });
+                              }else{
+                                popUpInfo(context, "Alert", "Error trying to delete notification. Only an Administrator can delete notifications");
+                              }
                             },
                           )
                         ],
